Configure `mx`. Refactor colors out of tailwind config.
