+++ packages/ui/react-ui-theme/src/config/tailwind.ts
@@ -2,12 +2,6 @@
 // Copyright 2022 DXOS.org
 //
 
-import {
-  curvePathFromPalette,
-  paletteShadesFromCurve,
-  Lab_to_hex as labToHex,
-  hex_to_LCH as hexToLch,
-} from '@fluent-blocks/colors';
 import tailwindcssForms from '@tailwindcss/forms';
 import merge from 'lodash.merge';
 import tailwindColors from 'tailwindcss/colors';
@@ -16,179 +10,11 @@ import { type Config, type ThemeConfig } from 'tailwindcss/types/config';
 import tailwindcssLogical from 'tailwindcss-logical';
 import tailwindcssRadix from 'tailwindcss-radix';
 
-import { semanticColors } from '../util/semanticColors';
+import { physicalColors, semanticColors } from './colors';
+import { semanticColors as semanticColorsPlugin } from '../util/semanticColors';
 
 export type TailwindConfig = Config;
 export type TailwindThemeConfig = ThemeConfig;
-
-export type PaletteConfig = {
-  keyColor: string;
-  darkCp: number;
-  lightCp: number;
-  hueTorsion: number;
-};
-
-const shadeNumbers: number[] = /* [...Array(19)].map((_, i) => 50 + i * 50); */ [
-  50, 100, 150, 200, 250, 300, 350, 400, 450, 500, 550, 600, 650, 700, 750, 800, 850, 900, 950,
-];
-
-const broadShadeNumbers: number[] = [
-  12, 25, 37, 50, 75, 100, 125, 150, 200, 250, 300, 350, 400, 450, 500, 550, 600, 650, 700, 750, 800, 825, 850, 875,
-  900, 925, 950, 975,
-];
-
-const dtor = Math.PI / 180;
-
-export const paletteConfigs: Record<string, PaletteConfig> = {
-  red: {
-    keyColor: '#ee003b',
-    darkCp: 1,
-    lightCp: 0.13,
-    hueTorsion: 5.5 * dtor,
-  },
-  orange: {
-    keyColor: '#fa6b32',
-    darkCp: 0.725,
-    lightCp: 1,
-    hueTorsion: 13 * dtor,
-  },
-  amber: {
-    keyColor: '#f08c00',
-    darkCp: 1,
-    lightCp: 1,
-    hueTorsion: 24 * dtor,
-  },
-  yellow: {
-    keyColor: '#ffd900',
-    darkCp: 1,
-    lightCp: 1,
-    hueTorsion: 32 * dtor,
-  },
-  lime: {
-    keyColor: '#99c400',
-    darkCp: 1,
-    lightCp: 1,
-    hueTorsion: -3.5 * dtor,
-  },
-  green: {
-    keyColor: '#30a908',
-    darkCp: 0.35,
-    lightCp: 0.665,
-    hueTorsion: -10.5 * dtor,
-  },
-  emerald: {
-    keyColor: '#15e066',
-    darkCp: 1,
-    lightCp: 0.735,
-    hueTorsion: -7 * dtor,
-  },
-  teal: {
-    keyColor: '#00a270',
-    darkCp: 1,
-    lightCp: 0.755,
-    hueTorsion: -12 * dtor,
-  },
-  cyan: {
-    keyColor: '#048992',
-    darkCp: 1,
-    lightCp: 0.855,
-    hueTorsion: -15 * dtor,
-  },
-  sky: {
-    keyColor: '#007bc2',
-    darkCp: 1,
-    lightCp: 0.64,
-    hueTorsion: 0,
-  },
-  blue: {
-    keyColor: '#0058cb',
-    darkCp: 1,
-    lightCp: 0.495,
-    hueTorsion: -7 * dtor,
-  },
-  indigo: {
-    keyColor: '#1b45c5',
-    darkCp: 0.495,
-    lightCp: 0.55,
-    hueTorsion: 0,
-  },
-  violet: {
-    keyColor: '#080886',
-    darkCp: 0.195,
-    lightCp: 0.635,
-    hueTorsion: -5 * dtor,
-  },
-  purple: {
-    keyColor: '#2c0073',
-    darkCp: 0,
-    lightCp: 0.63,
-    hueTorsion: 0,
-  },
-  fuchsia: {
-    keyColor: '#6d0085',
-    darkCp: 1,
-    lightCp: 0.695,
-    hueTorsion: 0,
-  },
-  pink: {
-    keyColor: '#a5006d',
-    darkCp: 1,
-    lightCp: 0.775,
-    hueTorsion: 0,
-  },
-  rose: {
-    keyColor: '#d00054',
-    darkCp: 1,
-    lightCp: 1,
-    hueTorsion: 0,
-  },
-  neutral: {
-    keyColor: '#707076',
-    darkCp: 0.8,
-    lightCp: 0.88,
-    hueTorsion: 0,
-  },
-  primary: {
-    keyColor: '#1662d9',
-    darkCp: 0.86,
-    lightCp: 1,
-    hueTorsion: -30 * dtor,
-  },
-};
-
-export const configColors = Object.keys(paletteConfigs).reduce(
-  (acc: Record<string, Record<string, string>>, palette) => {
-    const isBroad = palette === 'neutral' || palette === 'primary';
-    const paletteConfig = paletteConfigs[palette] as PaletteConfig;
-    const curve = curvePathFromPalette({
-      ...paletteConfig,
-      keyColor: hexToLch(paletteConfig.keyColor),
-    });
-    const defaultShades = paletteShadesFromCurve(curve, 21, [0, 100 * (22 / 21)], 1, 24).reverse();
-    const renderCssValue = (shadeNumber: number) => {
-      if (shadeNumber > 999) {
-        return '#000000';
-      } else if (shadeNumber < 1) {
-        return '#ffffff';
-      } else if (shadeNumber % 50 === 0) {
-        return labToHex(defaultShades[shadeNumber / 50]);
-      } else {
-        const k2 = (shadeNumber % 50) / 50;
-        const k1 = 1 - k2;
-        const [l1, a1, b1] = defaultShades[Math.floor(shadeNumber / 50)];
-        const [l2, a2, b2] = defaultShades[Math.ceil(shadeNumber / 50)];
-        return labToHex([l1 * k1 + l2 * k2, a1 * k1 + a2 * k2, b1 * k1 + b2 * k2]);
-      }
-    };
-    acc[palette] = (isBroad ? broadShadeNumbers : shadeNumbers).reduce((acc: Record<string, string>, shadeNumber) => {
-      acc[shadeNumber] = renderCssValue(shadeNumber);
-      return acc;
-    }, {});
-
-    return acc;
-  },
-  {},
-);
 
 export const tailwindConfig = ({
   env = 'production',
@@ -206,74 +32,6 @@ export const tailwindConfig = ({
       body: ['Inter Variable', ...defaultConfig.theme.fontFamily.sans],
       mono: ['JetBrains Mono Variable', ...defaultConfig.theme.fontFamily.mono],
     },
-    semanticColors: {
-      scrim: {
-        light: `${configColors.neutral['75']}A6`,
-        dark: `${configColors.neutral['900']}A6`,
-      },
-      attention: {
-        light: '#ffffff',
-        dark: configColors.neutral['900'],
-      },
-      base: {
-        light: configColors.neutral['25'],
-        dark: configColors.neutral['850'],
-        fg: { light: '#ffffff', dark: '#000000' },
-      },
-      description: {
-        light: configColors.neutral['300'],
-        dark: configColors.neutral['500'],
-      },
-      baseGlass: {
-        light: `${configColors.neutral['25']}E6`,
-        dark: `${configColors.neutral['850']}E6`,
-      },
-      input: {
-        light: configColors.neutral['50'],
-        dark: configColors.neutral['825'],
-      },
-      modal: {
-        light: '#ffffff',
-        dark: `${configColors.neutral['750']}`,
-      },
-      modalSelected: {
-        light: configColors.neutral['50'],
-        dark: configColors.neutral['825'],
-      },
-      hover: {
-        light: configColors.neutral['37'],
-        dark: configColors.neutral['800'],
-      },
-      accent: {
-        light: configColors.primary['550'],
-        dark: configColors.primary['550'],
-        fg: { light: configColors.primary['550'], dark: configColors.primary['400'] },
-      },
-      accentHover: {
-        light: configColors.primary['600'],
-        dark: configColors.primary['500'],
-      },
-      unAccent: {
-        light: configColors.neutral['500'],
-        dark: configColors.neutral['400'],
-      },
-      unAccentHover: {
-        light: configColors.neutral['400'],
-        dark: configColors.neutral['500'],
-      },
-      separator: {
-        light: configColors.neutral['75'],
-        dark: configColors.neutral['750'],
-      },
-      inverse: {
-        light: '#ffffff',
-        dark: '#ffffff',
-      },
-      unavailable: {
-        light: configColors.neutral['100'],
-        dark: configColors.neutral['600'],
-      },
-    },
     extend: merge(
       {
         screens: {
@@ -281,20 +39,21 @@ export const tailwindConfig = ({
           'hover-hover': { raw: '(hover: hover)' },
         },
         colors: {
-          ...configColors,
+          ...physicalColors,
           slate: tailwindColors.slate,
           gray: tailwindColors.gray,
           zinc: tailwindColors.zinc,
           stone: tailwindColors.stone,
-          success: configColors.emerald,
-          warning: configColors.amber,
-          error: configColors.rose,
-          info: configColors.cyan,
+          success: physicalColors.emerald,
+          warning: physicalColors.amber,
+          error: physicalColors.rose,
+          info: physicalColors.cyan,
           transparent: 'transparent',
           current: 'currentColor',
           white: '#ffffff',
           black: '#000000',
         },
+        semanticColors,
         fontSize: {
           // Base size 16px
           // Scale 1.125
@@ -415,7 +174,7 @@ export const tailwindConfig = ({
       ...extensions,
     ),
   },
-  plugins: [semanticColors, tailwindcssLogical, tailwindcssForms, tailwindcssRadix()],
+  plugins: [semanticColorsPlugin, tailwindcssLogical, tailwindcssForms, tailwindcssRadix()],
   ...(env === 'development' && { mode: 'jit' }),
   content,
   future: {
